+++ LinkEngine/GameTypes/RPG/Engine/Entities/Item.cs
@@ -2,6 +2,10 @@
 
 namespace LinkEngine.RPG
 {
+    /// <summary>
+    /// RPGItem is the the RPG derivitive of the Entities.Item Class.
+    /// It holds the cost of the item in ingame currency
+    /// </summary>
     public class RPGItem : Item
     {
         public int Cost { get; set; }
@@ -9,29 +13,6 @@ namespace LinkEngine.RPG
         public RPGItem (int id, string name, string namePlur, int cost) : base (id, name, namePlur)
         {
             Cost = cost;
-        }
-
-        public void Use(Character player)
-        {
-            Potion pot = (Potion)this;
-
-            switch (pot.VariableToBuff)
-            {
-                case "Health":
-                    player.Health += pot.AmountToBuff;
-                    break;
-                case "Strength":
-                    // player.Strength += pot.AmountToBuff;
-                    break;
-                case "Defense":
-                    // player.Defense += pot.AmountToBuff;
-                    break;
-
-            }
-        }
-        public void Equip()
-        {
-
         }
     }
 }